// src/app/tabs/shared/models/account-head.model.ts
export class AccountHead {
    rowNumber?: number;
    id?: number;
    billHeadId?: number;
    accountHeadName?: string;
    fixedAmount?: number;
    fixedAmountLocal?: string;
    discountAmount?: number;
    discountAmountLocal?: string;
    formulaFieldName?: string;
    operatorName?: string;
    formulaDescription?: string;
    operatorDescription?: string;
    position?: number;
}